attribute folder, add queryBased + nrt data models
