+++ .script/tests/detectionTemplateSchemaValidation/Models/ModelValidationAttributes/DictionaryKeyMatchesRegexAttribute.cs
