+++ .script/tests/detectionTemplateSchemaValidation/Models/ModelValidationAttributes/DictionaryValueMatchesRegexAttribute.cs
