+++ .script/tests/detectionTemplateSchemaValidation/Models/ModelValidationAttributes/FrequencyLimitationForLongPeriodQuery.cs
