+++ .script/tests/detectionTemplateSchemaValidation/Models/ModelValidationAttributes/NewEntityMappingsAttribute.cs
