+++ .script/tests/detectionTemplateSchemaValidation/Models/ModelValidationAttributes/QueryBasedTemplateVersionValidator.cs
