+++ .script/tests/detectionTemplateSchemaValidation/Models/ModelValidationAttributes/RangeTimeSpanIsoFormatAttribute.cs
