+++ .script/tests/detectionTemplateSchemaValidation/Models/ModelValidationAttributes/ValidEntityMappingsAttribute.cs
