+++ .script/tests/detectionTemplateSchemaValidation/Models/NrtTemplateInternalModel.cs
@@ -1,0 +1,8 @@
+
+
+namespace Microsoft.Azure.Sentinel.Analytics.Management.AnalyticsTemplatesService.Interface.Model
+{
+    public class NrtTemplateInternalModel : QueryBasedTemplateInternalModel
+    {
+    }
+}
